add state and onchange handlers for series ends element
